+++ src/execution/createPipelineExecutor.ts
@@ -1,11 +1,19 @@
 import { spaceTrim } from 'spacetrim';
 import type { Promisable } from 'type-fest';
-import { IS_VERBOSE, LOOP_LIMIT, MAX_EXECUTION_ATTEMPTS, MAX_PARALLEL_COUNT, RESERVED_PARAMETER_MISSING_VALUE, RESERVED_PARAMETER_NAMES, RESERVED_PARAMETER_RESTRICTED } from '../config';
+import {
+    IS_VERBOSE,
+    LOOP_LIMIT,
+    MAX_EXECUTION_ATTEMPTS,
+    MAX_PARALLEL_COUNT,
+    RESERVED_PARAMETER_MISSING_VALUE,
+    RESERVED_PARAMETER_NAMES,
+    RESERVED_PARAMETER_RESTRICTED,
+} from '../config';
 import { extractParameterNamesFromPromptTemplate } from '../conversion/utils/extractParameterNamesFromPromptTemplate';
 import { validatePipeline } from '../conversion/validation/validatePipeline';
+import { ExpectError } from '../errors/_ExpectError';
 import { PipelineExecutionError } from '../errors/PipelineExecutionError';
 import { UnexpectedError } from '../errors/UnexpectedError';
-import { ExpectError } from '../errors/_ExpectError';
 import { isValidJsonString } from '../formats/json/utils/isValidJsonString';
 import { joinLlmExecutionTools } from '../llm-providers/multiple/joinLlmExecutionTools';
 import { isPipelinePrepared } from '../prepare/isPipelinePrepared';
@@ -15,7 +23,13 @@ import type { PipelineJson } from '../types/PipelineJson/PipelineJson';
 import type { PromptTemplateJson } from '../types/PipelineJson/PromptTemplateJson';
 import type { ChatPrompt, CompletionPrompt, EmbeddingPrompt, Prompt } from '../types/Prompt';
 import type { TaskProgress } from '../types/TaskProgress';
-import type { Parameters, ReservedParameters, string_markdown, string_name, string_parameter_value } from '../types/typeAliases';
+import type {
+    Parameters,
+    ReservedParameters,
+    string_markdown,
+    string_name,
+    string_parameter_value,
+} from '../types/typeAliases';
 import { arrayableToArray } from '../utils/arrayableToArray';
 import type { really_any } from '../utils/organization/really_any';
 import type { TODO_any } from '../utils/organization/TODO_any';
@@ -400,55 +414,8 @@ export function createPipelineExecutor(options: CreatePipelineExecutorOptions): 
                                         ...currentTemplate.expectations,
                                     },
                                     expectFormat: currentTemplate.expectFormat,
-                                    postprocessing: (currentTemplate.postprocessingFunctionNames || []).map(
-                                        (functionName) => async (result: string) => {
-                                            // TODO: DRY [☯]
-                                            const errors: Array<Error> = [];
-                                            for (const scriptTools of arrayableToArray(tools.script)) {
-                                                try {
-                                                    return await scriptTools.execute({
-                                                        scriptLanguage: `javascript` /* <- TODO: Try it in each languages; In future allow postprocessing with arbitrary combination of languages to combine */,
-                                                        script: `${functionName}(result)`,
-                                                        parameters: {
-                                                            result: result || '',
-                                                            // Note: No ...parametersForTemplate, because working with result only
-                                                        },
-                                                    });
-                                                } catch (error) {
-                                                    if (!(error instanceof Error)) {
-                                                        throw error;
-                                                    }
-
-                                                    if (error instanceof UnexpectedError) {
-                                                        throw error;
-                                                    }
-
-                                                    errors.push(error);
-                                                }
-                                            }
-
-                                            if (errors.length === 0) {
-                                                throw new PipelineExecutionError(
-                                                    'Postprocessing in LlmExecutionTools failed because no ScriptExecutionTools were provided',
-                                                );
-                                            } else if (errors.length === 1) {
-                                                throw errors[0];
-                                            } else {
-                                                throw new PipelineExecutionError(
-                                                    spaceTrim(
-                                                        (block) => `
-                                                        Postprocessing in LlmExecutionTools failed ${errors.length}x
-
-                                                        ${block(
-                                                            errors.map((error) => '- ' + error.message).join('\n\n'),
-                                                        )}
-                                                      `,
-                                                    ),
-                                                );
-                                            }
-                                        },
-                                    ),
-                                } as Prompt;
+                                    postprocessingFunctionNames: currentTemplate.postprocessingFunctionNames,
+                                } as Prompt; // <- TODO: Not very good type guard
 
                                 variant: switch (currentTemplate.modelRequirements!.modelVariant) {
                                     case 'CHAT':
@@ -496,7 +463,6 @@ export function createPipelineExecutor(options: CreatePipelineExecutorOptions): 
                                 }
 
                                 // TODO: DRY [1]
-
                                 scriptPipelineExecutionErrors = [];
 
                                 // TODO: DRY [☯]